Generic Collection/Document reference class
